+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix NPC offers with thousand separators not being parsed properly (e.g. NPC Coco)
 - `Creature.mitigation` is now a float.
 - Also consider `Category:Unavailable` when skipping deprecated articles.
+- `Charm.cost` replaced by `cost_level_1`, `cost_level_2` and `cost_level_3` for each charm's level.
 - Renamed many item attributes (`item_attribute`) for more clarity:
   - `level` -> `required_level`
   - `vocation` -> `required_vocation`